Android fixes (#1614)
* Move initialization of `app` and `mainHandler` variables from onCreate to attachBaseContext

* Run onRestoreInstanceState earlier to stop it from undoing LiveData observer effects. For example:
Initiate opening a wallet and then immediately start a rotation or send the app to the background. The observer will notice the new wallet after the following onStart and show the TransactionsFragment. But then onRestoreInstanceState would restore the NoWalletFragment again.

* Add support links to About box  /  Use copyright statement from Qt app

* Preserve derivation text box enabled state across rotation

* Avoid race conditions by setting commands.wallet on foreground thread before closing a wallet and after opening one. For example:
When receiving a URI intent, it was possible that the check for an open wallet would succeed even though a background thread was in the middle of closing it. Then when the Send dialog attempted to access the wallet, the app would crash.

* Update to current stable androidx release, and use ktx extensions for LiveData.observe and Fragment.viewModels

* Avoid replacing SettingsFragment if it already exists (closes #1611)

* In NewWalletDialog1, preserve wallet type across rotation

* Use commitNow to fix TransactionFragment disappearing when switching between wallets

* Connect NewWalletDialog1 and NewWalletDialog1 using setTargetFragment

* Use new layoutId constructors on all activities and fragments

* Restructure LiveDatas to avoid redundant refreshes

* Update to Chaquopy 6.3.1 and Android Gradle plugin 3.5.0