iOS Fix: Fix for iOS 13.1.x bug where Touch ID prompt doesn't appear
This is a bug in iOS itself. See:

https://forums.developer.apple.com/thread/122628

Our workaround is to pop up a "please touch home button" prompt at the
same time as the built-in Touch ID prompt.  If the prompt from iOS
appears, it will cover up our prompt. If it doesn't, our prompt appears.

Note this bug has been confirmed fixed in iOS 13.2, so we only do this
workaround on iOS 13.1 and on phones with touchID (iPhone 6,7,8).
iPhoneX uses face ID.