iOS fixup: Wallet Seed dialog was sometimes not appearing
- When the iOS13.x Touch ID workaround was active, the Wallet Seed
dialog would sometimes not appear. This was because the "Please Use
TouchID" dialog was still up.  This was a latent bug anyway in
gui.get_presented_viewcontroller. Fixed.

- Updated the readme to warn about using Xcode 11.x or above to build
the app. Apple changed the iOS Viewcontroller API on us and we don't
have time to update the codebase right now. If built with Xcode 10.x,
everything works fine. If built with Xcode 11.x, everything broken.