iOS Status Bar workaround continued, made iPad on iOS13 always hide SB
Status bar workaround can't be done at this time for iPad (the situation
with iPad is exceedingly complex because it allows non-full-screen
windows of variable size).

So we just globally disable the app's status bar for iPad if on iOS 13.

On iPhoneX the workaround is not needed.

On iPhone non-X with iOS 13 the workaround is needed and enabled.