iOS fix: "Downloading headers xx%..." is now more accurate
The code for computing the percentage is old and pre-dated our EX
protocol 1.4 updates.  It calculated the percentage as a total of the
entire blockchain.

We updated the code to be a percentage work of the *current* download.
Eg if you have header 12345 and the blockchain is at 22445, you start at
0% and work your way up to 100% at 22445.  (Before it would show 99%...
for the whole time).

Also in this commit: remove dead code for ios/