Daemon/Lib: register lockfile removal atexit.
This should fix the issue on Linux where sometimes the jsonrpc client
lib goes out to lunch truying to contact the daemon.

See: https://github.com/joshmarshall/jsonrpclib/issues/12

Issue since 2014.  It still happens on latest release.

By deleting the lockfile at app exit when we are the ones that created
it, we guarantee that it's not around the next time the app is run and
this should reduce the possibility of the "jsonrpc client out to lunch"
situation some users are reporting.

The real fix is to monkey-patch jsonrpclib to specify a timeout on the
socket -- but that requires further investigation so we do this for now.