Plugins: Fix to @hook / @run_hook mechanism (#1669)
* Fix to plugin run_hook mechanism

The @hook decorator was registering functions globally for all plugins,
even ones not decorated with @hook would get picked up.

This change fixes it so that only functions decorated with @hook get
picked up.

This commit is not 100% ready for merge to mainline.  I still have to do
some due diligence and verify no extant plugins DEPENDED on the old
behavior for correct operation!

* Follow-up to @hook fix: Deal with @property attributes properly

What we were doing before could trigger an @property instance property
to execute a function as part of our getattr() check in _is_hook.

This commit handles the property situation correctly.

* @hook fix follow-up: Check for _is_ec_plugin_hook on class-level

This may be faster, and was suggested by Mark.

* Nit: Allow run_hook hooks to return more than 1 result

We just pick the first one and print a warning to the console.

* plugins.py nit: Alphabeticized the imports

My OCD is satisfied.