Fix #1671 -- No FX data in CLI 'history' even if --show_fiat
There was a bug in the code. It needed to invoke fx.run() at least once
in order to ensure the history data is loaded from disk cache and/or
perhaps even downloaded in the background.

Ideally we would wait for the fx thread to have downloaded data -- but
this fix at least ensures cached data is returned right away.