Follow-up: Suppresses OP_RETURN & scripts when exporting history to CSV
Exporting history in the GUI produced a CSV file that looked very
unuseful in the presence of scripts/OP_RETURNs.

This commit fixes that.  Only valid addresses in a tx are exported to CSV.
Scripts are suppressed from CSV export in the "addresses" columns.

Exporting to JSON, however, does export all addresses parsed (even if
script).  Presumably if using JSON as a data source, user can deal with
the weirdness of scripts.

Whereas CSV is for Excel, etc.. so it's better to leave out scripts and
only include addresses.