Update burn tool for multisig signing
This commit allows users having multisig wallets to import an unsigned transaction hex /or file being used in a burn transaction.  Before the user is allowed to sign the transaction he must enter the proper amount of tokens into the dialog's burn quanity field.

Also the "amt_to_burn" property was added to the "check_tx_slp" method for the pre-signing burn prevention check.  This requires the user to specify both the coins to be used in the burn transaction and also the proper amount to be burned.