Prompt user if token id is incorrect
If transaction to burn contains an SLP message, then check the token ID matches.  If there is no SLP message in the burn transaction and the amount specified by the user is mismatched, then tell the user to double check the token ID of the burn tool dialog matches the id of the token that should be burned.