Satochip hotfix: Fix issue #1734
The actual bug is not fixed -- merely the lib code that triggers the
exception has been patched to be more resilient.

The core problem remains but from a user's perspective it "looks fixed".

The core problem really is the client code somehow is spamming
exceptions and then killing the parent windows that are showing the
error dialogs.. or something like that.