Fix# 1707: Suppress main_winodw.pop_top_level_window ValueError
It's useful to be able to call this function multiple times from the
same window or when the window never got a chance to get "pushed" onto
the stack (e.g. from an unconditional cleanup handler).

Some code in transaction_dialog was already doing this, thus causing an
exception in pathological cases.

We made this function more resilient and more useful for callers by
suppressing ValueError if the window is not in the stack.

Optionally if the caller cares about this Exception (as is the case in
some codepaths already in the codebase) -- they may specify
'raise_if_missing=True' to pop_top_level_window.

Closes #1707.