initial sketch for exclude_txids, still needs...
TODO: currently the validator is setup to assume any missing txn from the gs response is invalid SLP.  This allows us to speed up DAG crawl significantly since we don't request these invalid txns.  However,  this needs to be modified to first check if the txid is valid in cache before marking as invalid automatically.